Restructurisation of Handlers
