+++ src/Main.java
@@ -8,6 +8,8 @@
  *
  * @author Anton Sasnouski
  */
+import MainFunctions.DataManager;
+import MainFunctions.Handlers;
 import org.telegram.telegrambots.meta.TelegramBotsApi;
 import org.telegram.telegrambots.meta.exceptions.TelegramApiException;
 import org.telegram.telegrambots.updatesreceivers.DefaultBotSession;
