+++ src/MainFunctions/Config.java
@@ -1,5 +1,7 @@
+package MainFunctions;
+
 public class Config {
     public static final String TOKEN = "";
     public static final String STEAM_KEY = "";
-    public static final long TG_ID = 111111111;
+    public static final long TG_ID = 111111;
 }
